Move Base64 builders to ...encoding.base64 pkg + deprecate ...encoding.builders pkg classes/functions
